pyproject.toml
[project]
name = "speech-emotion-recognition"
version = "0.1.0"
description = "Classifier of 1 of 6 emotions in audio segment"
authors = [
    {name = "IvanZhigalskii"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "torchaudio (>=2.6.0,<3.0.0)",
    "fire (>=0.7.0,<0.8.0)",
    "lightning (>=2.5.0.post0,<3.0.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "omegaconf (>=2.3.0,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "kagglehub (>=0.3.10,<0.4.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "torch-audiomentations (>=0.12.0,<0.13.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"
dvc = "^3.59.1"
pre-commit = "^4.2.0"
pytest = "^8.3.5"
pytest-cov = "^6.1.0"
wandb = "^0.19.8"


[[tool.poetry.source]]
name = "torch-cuda"
url = "https://download.pytorch.org/whl/cu126"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "torch-cuda"}
torchaudio = {source = "torch-cuda"}

[tool.isort]
profile = "black"
